Added rudimentary tournament backbone Fixed jump off board bug Fixed KingMe jump bug
-- Inspect Stall on WHITE victory

+++ Board.h
@@ -10,7 +10,9 @@
 #ifndef FILE_BOARD_H_INCLUDED
 #define FILE_BOARD_H_INCLUDED
 
-typedef unsigned long	ULONG;
+#include <stdint.h>
+typedef uint32_t	ULONG;
+
 
 class board{
 public:
